requirements.txt
# Runtime (Python 3.11+)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-telegram-bot>=20.0
openai>=1.40.0
gspread>=6.0.0
google-auth>=2.29.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
ruff>=0.6.0
